Expand Pool Page Size to 64KiB and Counters to int32
Increase block/scalable pool page size from 4 KiB to 64 KiB and migrate related address, size, and usage fields from int16 to int32. This updates free-list bookkeeping, double-allocation/free tracking, pointer offset casts, and usage percentile return types so larger pages are handled safely without 16-bit overflow. Capacity-selection assertions were also tightened to validate non-negative requests against the updated limits.

+++ SDK/Core/Include/FE/pool/private/pool_common.hxx
@@ -69,8 +69,8 @@ namespace internal::pool
 
     struct block_info
     {
-        var::int16 _address;
-        var::int16 _size_in_bytes;
+        var::int32 _address;
+        var::int32 _size_in_bytes;
     };
 
     class from_low_address
